requirements.txt
addict==2.4.0
aiofiles==23.2.1
albucore==0.0.23
albumentations==2.0.4
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
crcmod==1.7
cryptography==44.0.3
cycler==0.12.1
dill==0.4.0
einops==0.8.1
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gradio==4.44.1
gradio_client==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jmespath==0.10.0
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.14.0
onnxruntime==1.15.1
onnxruntime-gpu==1.18.0
onnxslim==0.1.31
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
orjson==3.10.18
oss2==2.17.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.30.2
psutil==5.9.8
py-cpuinfo==9.0.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-wavelets==1.3.0
pytz==2023.4
PyWavelets==1.8.0
PyYAML==6.0.2
requests==2.28.2
rich==13.4.2
ruff==0.11.8
safetensors==0.5.3
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.13.1
tabulate==0.9.0
termcolor==3.1.0
timm==0.6.7
tomlkit==0.12.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.65.2
triton==3.2.0
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
# Editable install with no version control (ultralytics==8.3.13)
-e /root/ultralyticsPro-YOLOv12
ultralytics-thop==2.0.14
urllib3==1.26.20
uvicorn==0.34.2
websockets==12.0
yapf==0.43.0
